pyproject.toml
[project]
name = "moments_fsd"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "numpy",
    "scipy",
    "matplotlib",
    "tqdm",
    "wandb",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
